feat: implement syncDocNodeToLexical functionality
- Introduced the `syncDocNodeToLexical` function to synchronize document changes with the Lexical editor, enabling real-time updates during collaborative editing.
- Added `transformSelection` utility to manage selection updates based on document changes, ensuring accurate cursor positioning and selection states.
- Organized the `syncDocNodeToLexical` module into a dedicated directory for better structure and maintainability.
- Updated tests to validate the new synchronization features and selection transformations, enhancing the overall collaborative editing experience.
